feat: Add POST endpoint for creating users
This commit adds a new POST endpoint for creating users in the userRoute.js file. The endpoint is responsible for creating a new user in the database with the provided name and password. It performs validation to ensure that both name and password are provided in the request body. If the user already exists in the database, a 409 conflict response is returned. Otherwise, the user is successfully created and a 201 created response is returned with the newly created user object.

Note: This commit message follows the established convention of starting with a type (feat) followed by a brief description of the changes.